+++ lib/px/utils/pxcookiev1.lua
@@ -38,8 +38,7 @@ function PXCookieV1:validate(data)
         self.px_logger.debug('cookie verification succeed with no IP in signature')
         return true
     end
-
-    self.px_logger.error('Failed to verify cookie v1 content ' .. self.cjson.encode(data));
+    self.px_logger.debug('Cookie HMAC validation failed, value without ip: '.. digest_noip ..' with ip: '.. digest_ip ..', user-agent: ' .. ngx.req.get_headers()["User-Agent"]);
     return false
 end
 
@@ -55,14 +54,14 @@ function PXCookieV1:process()
         local success, result = pcall(self.decrypt, self, cookie, self.cookie_secret)
 
         if not success then
-            self.px_logger.error("Could not decrpyt cookie - " .. result)
+            self.px_logger.debug("Could not decrpyt cookie - " .. result)
             error({ message = "cookie_decryption_failed" })
         end
         data = result["plaintext"]
     else
         local success, result = pcall(ngx.decode_base64, cookie)
         if not success then
-            self.px_logger.error("Could not decode b64 cookie - " .. result)
+            self.px_logger.debug("Could not decode b64 cookie - " .. result)
             error({ message = "cookie_decryption_failed" })
         end
         data = result
@@ -71,7 +70,7 @@ function PXCookieV1:process()
     -- Deserialize the JSON payload
     local success, result = pcall(self.decode, self, data)
     if not success then
-        self.px_logger.error("Could not decode cookie")
+        self.px_logger.debug("Could not decode cookie")
         error({ message = "cookie_decryption_failed" })
     end
 
@@ -89,24 +88,30 @@ function PXCookieV1:process()
 
     -- cookie expired
     if fields.t and fields.t > 0 and fields.t / 1000 < os.time() then
-        self.px_logger.debug("Cookie expired - " .. data)
+        self.px_logger.debug('Cookie TTL is expired, value: '.. data ..', age: ' .. fields.t / 1000 - os.time())
         error({ message = "cookie_expired" })
     end
 
     -- Set the score header for upstream applications
     self.px_headers.set_score_header(fields.s.b)
+    -- Set the score variable for logging 
+    self.px_logger.set_score_variable(fields.s.b)
+    
     -- Check bot score and block if it is >= to the configured block score
-    if fields.s and fields.s.b and fields.s.b >= self.blocking_score then
+    if fields.s and fields.s.b then
         self.px_logger.debug("Visitor score is higher than allowed threshold: " .. fields.s.b)
         ngx.ctx.px_action = 'c'
         ngx.ctx.block_score = fields.s.b
+    end
+
+    if fields.s.b >= self.blocking_score then
         return false
     end
 
     -- Validate the cookie integrity
     local success, result = pcall(self.validate, self, fields)
     if not success or result == false then
-        self.px_logger.error("Could not validate cookie v1 signature - " .. data)
+        self.px_logger.debug("Could not validate cookie v1 signature - " .. data)
         error({ message = "cookie_validation_failed" })
     end