+++ lib/px/utils/pxcookiev3.lua
@@ -18,8 +18,7 @@ function PXCookieV3:validate(data)
     if digest == string.upper(ngx.ctx.px_cookie_hmac) then
         return true
     end
-
-    self.px_logger.error('Failed to verify cookie v3 content ' .. data);
+    self.px_logger.debug('Cookie HMAC validation failed, hmac: '.. digest ..', user-agent: ' .. ngx.req.get_headers()["User-Agent"]);
     return false
 end
 
@@ -35,7 +34,7 @@ function PXCookieV3:process()
         -- self:decrypt(cookie, self.cookie_secret)
         local success, result = pcall(self.decrypt, self, cookie, self.cookie_secret)
         if not success then
-            self.px_logger.error("Could not decrpyt px cookie v3")
+            self.px_logger.debug("Could not decrpyt px cookie v3")
             error({ message = "cookie_decryption_failed" })
         end
         data = result['plaintext']
@@ -44,7 +43,7 @@ function PXCookieV3:process()
         hash, orig_cookie = self:split_decoded_cookie(cookie);
         local success, result = pcall(ngx.decode_base64, orig_cookie)
         if not success then
-            self.px_logger.error("Could not decode b64 cookie - " .. result)
+            self.px_logger.debug("Could not decode b64 cookie - " .. result)
             error({ message = "cookie_decryption_failed" })
         end
         data = result
@@ -53,7 +52,7 @@ function PXCookieV3:process()
     -- Deserialize the JSON payload
     local success, result = pcall(self.decode, self, data)
     if not success then
-        self.px_logger.error("Could not decode cookie")
+        self.px_logger.debug("Could not decode cookie")
         error({ message = "cookie_decryption_failed" })
     end
 
@@ -71,24 +70,30 @@ function PXCookieV3:process()
 
     -- cookie expired
     if fields.t and fields.t > 0 and fields.t / 1000 < os.time() then
-        self.px_logger.debug("Cookie expired - " .. data)
+        self.px_logger.debug('Cookie TTL is expired, value: '.. data ..', age: ' .. fields.t / 1000 - os.time())
         error({ message = "cookie_expired" })
     end
 
     -- Set the score header for upstream applications
     self.px_headers.set_score_header(fields.s)
+    -- Set the score variable for logging
+    self.px_logger.set_score_variable(fields.s)
+    
     -- Check bot score and block if it is >= to the configured block score
-    if fields.s and fields.s >= self.blocking_score then
-        self.px_logger.debug("Visitor score is higher than allowed threshold: " .. fields.s)
+    if fields.s and fields.a then
         ngx.ctx.block_score = fields.s
         ngx.ctx.block_action = fields.a
+    end
+
+    if fields.s >= self.blocking_score then
+        self.px_logger.debug("Visitor score is higher than allowed threshold: " .. fields.s)
         return false
     end
 
-    -- Validate the cookie integrity
+        -- Validate the cookie integrity
     local success, result = pcall(self.validate, self, orig_cookie)
     if success == false or result == false then
-        self.px_logger.error("Could not validate cookie v3 signature - " .. orig_cookie)
+        self.px_logger.debug("Could not validate cookie v3 signature - " .. orig_cookie)
         error({ message = "cookie_validation_failed" })
     end