+++ lib/px/utils/pxtokenv1.lua
@@ -39,35 +39,25 @@ function TokenV1:validate(data)
         return true
     end
 
-    self.px_logger.error('Failed to verify cookie v1 content ' .. cjson.encode(data));
+    self.px_logger.debug('Cookie HMAC validation failed, value without ip: '.. digest_noip ..' with ip: '.. digest_ip ..', user-agent: ' .. ngx.req.get_headers()["User-Agent"]);
     return false
 end
 
 function TokenV1:process()
     local cookie = ngx.ctx.px_orig_cookie
-    if not cookie or cookie == "1" then
-        self.px_logger.error("no token available")
-        error({ message = "no_cookie" })
-    end
-
-    if cookie == "2" then
-        self.px_logger.error("mobile sdk was unable to reach the server ")
-        error({ message = "mobile_sdk_connection_error" })
-    end
-
     -- Decrypt AES-256 or base64 decode cookie
     local data
     if self.cookie_encrypted == true then
-        local success, result = pcall(self.decrypt, self, cookie, self.cookie_secret)
+        local success, result = pcall(self.pre_decrypt, self, cookie, self.cookie_secret)
         if not success then
-            self.px_logger.error("Could not decrpyt cookie - " .. result)
-            error({ message = "cookie_decryption_failed" })
+            self.px_logger.debug("Could not decrpyt cookie - " .. result["message"])
+            error({ message = result["message"] })
         end
         data = result["plaintext"]
     else
         local success, result = pcall(ngx.decode_base64, cookie)
         if not success then
-            self.px_logger.error("Could not decode b64 cookie - " .. result)
+            self.px_logger.debug("Could not decode b64 cookie - " .. result)
             error({ message = "cookie_decryption_failed" })
         end
         data = result
@@ -76,7 +66,7 @@ function TokenV1:process()
     -- Deserialize the JSON payload
     local success, result = pcall(self.decode, self, data)
     if not success then
-        self.px_logger.error("Could not decode cookie")
+        self.px_logger.debug("Could not decode cookie")
         error({ message = "cookie_decryption_failed" })
     end
 
@@ -94,24 +84,30 @@ function TokenV1:process()
 
     -- cookie expired
     if fields.t and fields.t > 0 and fields.t / 1000 < os.time() then
-        self.px_logger.debug("Cookie expired - " .. data)
+        self.px_logger.debug('Cookie TTL is expired, value: '.. data ..', age: ' .. fields.t / 1000 - os.time())
         error({ message = "cookie_expired" })
     end
 
     -- Set the score header for upstream applications
     self.px_headers.set_score_header(fields.s.b)
+    -- Set the score variable for logging
+    self.px_logger.set_score_variable(fields.s.b)
+    
     -- Check bot score and block if it is >= to the configured block score
-    if fields.s and fields.s.b and fields.s.b >= self.blocking_score then
+    if fields.s then
         self.px_logger.debug("Visitor score is higher than allowed threshold: " .. fields.s.b)
         ngx.ctx.px_action = 'c'
         ngx.ctx.block_score = fields.s.b
+    end
+
+    if fields.s.b and fields.s.b >= self.blocking_score then
         return false
     end
 
     -- Validate the cookie integrity
     local success, result = pcall(self.validate, self, fields)
     if not success or result == false then
-        self.px_logger.error("Could not validate cookie v1 signature - " .. data)
+        self.px_logger.debug("Could not validate cookie v1 signature - " .. data)
         error({ message = "cookie_validation_failed" })
     end