+++ t/004-expiredCookie.t
@@ -110,4 +110,5 @@ User-Agent:  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 
 --- error_code: 200
 
 --- error_log
-PX DEBUG: Cookie expired 
+[PerimeterX - DEBUG] [ PX_APP_ID ] - Cookie TTL is expired
+