+++ t/014-scoreVariable.t
@@ -6,35 +6,25 @@ sub bake_cookie {
     use Crypt::Mac::HMAC 'hmac_hex';
     use Crypt::Mode::CBC;
 
-    my ( $ip, $score, $uuid, $vid, $time ) = @_;
-    my $data = $time . '0' . $score . $uuid . $vid . $ip;
+    my ( $ip, $ua, $score, $uuid, $vid, $time ) = @_;
 
     my $password        = "perimeterx";
     my $salt            = '12345678123456781234567812345678';
-    my $iteration_count = 1000;
+    my $iteration_count = 1000; 
     my $hash_name       = undef;                              #default is SHA256
     my $len             = 48;
 
     my $km = pbkdf2( $password, $salt, $iteration_count, $hash_name, $len );
     my $key = substr( $km, 0,  32 );
     my $iv  = substr( $km, 32, 48 );
-
+    my $action = 'a';
     my $m         = Crypt::Mode::CBC->new('AES');
-    my $hmac      = hmac_hex( 'SHA256', $password, $data );
-    my $plaintext = '{"t":'
-      . $time
-      . ', "s":{"b":'
-      . $score
-      . ', "a":0}, "u":"'
-      . $uuid
-      . '", "v":"'
-      . $vid
-      . '", "h":"'
-      . $hmac . '"}';
+    my $plaintext = '{"u":"' . $uuid. '", "v":"' . $vid . '", "t":' . $time . ', "s":'. $score . ', "a":"' . $action . '"}';
     my $ciphertext = $m->encrypt( $plaintext, $key, $iv );
-
     my $cookie = encode_b64($salt) . ":" . 1000 . ":" . encode_b64($ciphertext);
-    return 'X-PX-AUTHORIZATION: 1:' . $cookie;
+    my $hmac      = hmac_hex( 'SHA256', $password, $cookie . $ua );
+    $cookie = $hmac . ":" . $cookie;
+    return 'Cookie: _px3=' . $cookie;
 }
 
 add_block_preprocessor(
@@ -43,7 +33,8 @@ add_block_preprocessor(
         my $time   = ( time() + 360 ) * 1000;
         my $cookie = bake_cookie(
             "1.2.3.4",
-            '0',
+'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.84 Safari/537.36',
+            '100',
             "57ecdc10-0e97-11e6-80b6-095df820282c",
             "vid",
             $time
@@ -60,10 +51,20 @@ run_tests();
 __DATA__
 
 
-=== TEST 1: Procoss Workflow
-Process a valid V1 cookie
+=== TEST 1: Test Score Header
+Set the NGX $pxscore variable
 
 --- http_config
+    map score $pxscore {
+        default 'nil';
+    }
+    
+    log_format enriched '$remote_addr - $remote_user [$time_local] '
+                    '"$request" $status $body_bytes_sent '
+                    '"$http_referer" "$http_user_agent" perimeterx_score "$pxscore';
+
+    access_log /var/log/nginx/access_log enriched;
+
     lua_package_path "/usr/local/lib/lua/?.lua;/usr/local/openresty/lualib/?.lua;;";
     lua_ssl_trusted_certificate "/etc/ssl/certs/ca-certificates.crt";
     lua_ssl_verify_depth 3;
@@ -79,21 +80,21 @@ Process a valid V1 cookie
     location = /t {
         resolver 8.8.8.8;
         set_by_lua_block $config {
-            pxconfig = require "px.pxconfig"
-            pxconfig.cookie_secret = "perimeterx"
-            pxconfig.px_debug = true
-            pxconfig.block_enabled = true
-            pxconfig.send_page_requested_activity = false
+    	    pxconfig = require "px.pxconfig"
+    	    pxconfig.cookie_secret = "perimeterx"
+    	    pxconfig.px_debug = true
+    	    pxconfig.block_enabled = false
+    	    pxconfig.send_page_requested_activity = false
             pxconfig.enable_server_calls  = false
             return true
         }
 
-        access_by_lua_block {
-        require("px.pxnginx").application()
-    }
+    	access_by_lua_block {
+	    require("px.pxnginx").application()
+	}
 
         content_by_lua_block {
-             ngx.say(ngx.var.remote_addr)
+             ngx.say(ngx.var.pxscore)
         }
 }
 
@@ -106,10 +107,4 @@ X-Forwarded-For: 1.2.3.4
 User-Agent:  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_4) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.84 Safari/537.36
 
 --- response_body
-1.2.3.4
-
---- error_code: 200
-
---- error_log
-[PerimeterX - DEBUG] [ PX_APP_ID ] - Cookie evaluation ended successfully, risk score: 0
-
+100